+++ src/controller/questionController/index.ts
@@ -1,0 +1,4 @@
+import { questionUseCase } from "../../usecases/questionUseCase";
+import { QuestionController } from "./questionController";
+
+export const questionController = new QuestionController(questionUseCase)
