--- src/controller/sendInformationController/index.ts
@@ -1,4 +1,0 @@
-import { sendInformationUseCase } from "../../usecases/sendInformationUseCase";
-import { SendInformationController } from "./sendInformation";
-
-export const sendInformationController = new SendInformationController(sendInformationUseCase)
