+++ src/usecases/questionUseCase/index.ts
@@ -1,0 +1,3 @@
+import { QuestionUseCase } from "./questionUseCase";
+
+export const questionUseCase = new QuestionUseCase()
