--- src/usecases/sendInformationUseCase/index.ts
@@ -1,3 +1,0 @@
-import { SendInformationUseCase } from "./sendInformationUseCase";
-
-export const sendInformationUseCase = new SendInformationUseCase()
